% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_count.R
\name{count_cellsize}
\alias{count_cellsize}
\title{Given a individual level population dataset and variables of interest, count the number of
population and the fraction of that subpopulationin the geography of interest}
\usage{
count_cellsize(size, popvar, geovar, ...)
}
\arguments{
\item{size}{A population, or census data frame}

\item{popvar}{The variable that indicates the population}

\item{geovar}{The variable that indicates the final grouping variable of interest}

\item{...}{variables that are in the regression}
}
\description{
Given a individual level population dataset and variables of interest, count the number of
population and the fraction of that subpopulationin the geography of interest
}
\examples{
size <- read_csv("https://www.shirokuriwaki.com/datasets/popsize08.csv")
size_cell <- count_cellsize(size, popvar = pop2008, geovar = stt, eth, age, sex, inc, edu)

}
