% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{eval_mrp}
\alias{eval_mrp}
\title{function to evaluate RMSE and correlation}
\usage{
eval_mrp(tbl, truth, est_suffix)
}
\arguments{
\item{truth}{variable name of the truth to predict (no quotes)}

\item{est_suffix}{regex for variable names of estimates to consider}
}
\value{
A dataframe with three columns, \code{model}, \code{rmse}, and \code{corr}.
One row for each model specified
}
\description{
function to evaluate RMSE and correlation
}
