% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat.R
\name{mrp_estimate}
\alias{mrp_estimate}
\title{wrapper function for MRP estimates}
\usage{
mrp_estimate(
  model,
  est_name,
  cells = size_cell,
  geovar = stt,
  fracvar = frac_of_state,
  return_vec = FALSE
)
}
\arguments{
\item{model}{a lm, glm, glmer, or sparsereg object}

\item{est_name}{pick a name for your estimate}

\item{cells}{cell counts produced from count_cellsize}

\item{geovar}{variable for geography of interest}

\item{fracvar}{variabel for the cell fraction}

\item{return_vec}{whether to return the results as a data frame or vector}
}
\description{
wrapper function for MRP estimates
}
