% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{scatter_mrp}
\alias{scatter_mrp}
\title{Diagnostic graphic}
\usage{
scatter_mrp(x, y, ggtemplate = gg0, data = mrp_df,
  null = global_mean)
}
\arguments{
\item{x}{the xaxis variable (no quotes)}

\item{y}{the yaxis variable (no quotes)}

\item{ggtemplate}{an empty ggplot object with formatting layout}

\item{data}{a dataframe which includes columns x and y}

\item{null}{the global mean a null benchmark}
}
\description{
Scatterplot with x as truth and y as estiamte
}
