% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_mrp_data}
\alias{simulate_mrp_data}
\title{Simulate biased survey samples}
\usage{
simulate_mrp_data(n)
}
\arguments{
\item{n}{Total survey sample to draw}
}
\description{
This code is taken from https://github.com/lauken13/rstanarm, and
designed by the three authors listed above.
}
\author{
Lauren A. Kennedy, Jonah Gabry, and Aki Vehtari
}
