% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststrat.R
\name{sum_to_geo}
\alias{sum_to_geo}
\title{Sum up the number of observations based on a geography}
\usage{
sum_to_geo(data, geovar, predvar, fracvar, name)
}
\arguments{
\item{data}{the dataset that has all the variables}

\item{geovar}{the variable indicating the geography to sum up to}

\item{predvar}{the variable for the predicted value}

\item{fracvar}{the variable that indicates the relative size of the cell, a string}

\item{name}{name for the new var, a string}
}
\description{
Sum up the number of observations based on a geography
}
