% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synth.R
\name{synth_cellsize}
\alias{synth_cellsize}
\title{Synthetic expansion}
\usage{
synth_cellsize(start, marginal)
}
\arguments{
\item{start}{starting cell counts. These will have been estimated by census data.}

\item{marginal}{new information we want to incorporate by synthetic approach,
usually only as a marginal distribution. Must be a dataframe with three variables:
the geography variable (same name for original count data), one variable that indicates the
category of the new variable (e.g. "voted" and "not voted"), and a variable called "frac" which
indicates the proportion of people of that category within that geography --
i.e. all values within a geography must sum to 1.}
}
\description{
Creates a synthetic post-stratification dataset by simple multiplication, i.e.
 distribution of variables in new cells (entered as marginals) are orthogonal to
 original counts
}
