% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{unit_rmse}
\alias{unit_rmse}
\title{RMSE between two vectors}
\usage{
unit_rmse(mu = truth, muhat)
}
\arguments{
\item{mu}{A vector, conventionally the estimand}

\item{muhat}{same length vectorm conventionally the estimate}
}
\description{
RMSE between two vectors
}
