% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{dev_logit}
\alias{dev_logit}
\title{Deviance from predicted values of a logit model}
\usage{
dev_logit(y, xb = NULL, phat = NULL, w = 1)
}
\arguments{
\item{y}{vector of the outcome that was regressed on, or the true values to predict}

\item{xb}{estimates from the logit model, before the link function is applied.}

\item{phat}{estimates from the logit model, after link function is applied.
Only either \code{xb} or \code{phat} is necessary.}

\item{w}{survey weights if appplicable}
}
\description{
Deviance from predicted values of a logit model
}
\details{
The deviance without weights, \eqn{dev(\hat p)} for target \eqn{y} is computed as

\deqn{-2 \times \sum_{i=1}^n y_i \log(\hat{p_i}) + (1 - y_i) \log(1 - \hat{p_i})}

where

\deqn{\hat{p} = \frac{\exp{X\beta}}{\exp{X\beta} + 1}}.

With weights, this simply becomes

 \deqn{-2 \times \sum_{i=1}^n w_i y_i \log(\hat{p_i}) + w_i (1 - y_i) \log(1 - \hat{p_i})}
}
