% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{dev_logit}
\alias{dev_logit}
\title{Deviance from a logit model}
\usage{
dev_logit(y, xb, w = 1)
}
\arguments{
\item{y}{vector of the outcome that was regressed on}

\item{xb}{estimates from the logit coefficient}

\item{w}{survey weights if appplicable}
}
\description{
Deviance from a logit model
}
